#9 #15 임베딩
