--- NSC/Tensorflow_Test/settings.py
@@ -1,24 +1,0 @@
-import locale
-import os
-import platform
-import datetime
-import time
-
-FORMAT_DATE = "%Y%m%d"
-FORMAT_DATETIME = "%Y%m%d%H%M%S"
-
-# 경로 설정
-BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-
-
-# 로케일 설정
-if 'Linux' in platform.system() or 'Darwin' in platform.system():
-    locale.setlocale(locale.LC_ALL, 'ko_KR.UTF-8')
-elif 'Windows' in platform.system():
-    locale.setlocale(locale.LC_ALL, '')
-
-def get_time_str():
-    return datetime.datetime.fromtimestamp(
-        int(time.time())).strftime(FORMAT_DATETIME)
-
-timestr = get_time_str()
