--- NSC/Tensorflow_Test/utils.py
@@ -1,24 +1,0 @@
-import time
-import datetime
-import numpy as np
-
-
-# 날짜, 시간 관련 문자열 형식
-FORMAT_DATE = "%Y%m%d"
-FORMAT_DATETIME = "%Y%m%d%H%M%S"
-
-
-def get_today_str():
-    today = datetime.datetime.combine(
-        datetime.date.today(), datetime.datetime.min.time())
-    today_str = today.strftime('%Y%m%d')
-    return today_str
-
-
-def get_time_str():
-    return datetime.datetime.fromtimestamp(
-        int(time.time())).strftime(FORMAT_DATETIME)
-
-
-def sigmoid(x):
-    return 1. / (1. + np.exp(-x))
