+++ NSC/environment.py
@@ -1,0 +1,28 @@
+class Environment:
+    PRICE_IDX = 4  #종가의 위치
+
+#생성자에게 관리할 차트 데이터 할당.
+    #차트 데이터는 2차원 배열이라고 생각하면 됨.
+    def __init__(self, chart_data=None):
+        self.chart_data = chart_data
+        self.observation = None
+        self.idx = -1
+
+    def reset(self):
+        self.observation = None
+        self.idx = -1
+
+    def observe(self):
+        if len(self.chart_data) > self.idx + 1:
+            self.idx += 1
+            self.observation = self.chart_data.iloc[self.idx]
+            return self.observation
+        return None
+
+    def get_price(self):
+        if self.observation is not None:
+            return self.observation[self.PRICE_IDX]
+        return None
+
+    def set_chart_data(self, chart_data):
+        self.chart_data = chart_data
