+++ NSC/policy_network.py
@@ -1,0 +1,44 @@
+import numpy as np
+from tensorflow.keras.models import Sequential
+from tensorflow.keras.layers import Activation, LSTM, Dense, BatchNormalization
+from tensorflow.keras.optimizers import SGD
+class PolicyNetwork:
+    def __init__(self, input_dim=0, output_dim=0, lr=0.01):
+        self.input_dim = input_dim
+        self.lr = lr
+
+        #LSTM 신경망
+        self.model = Sequential()
+
+        self.model.add(LSTM(256,input_shape=(1, input_dim),
+                            return_sequences=True, stateful=False, dropout=0.5))
+        self.model.add(BatchNormalization())
+        self.model.add(LSTM(256,return_sequences=True, stateful=False, dropout=0.5))
+        self.model.add(BatchNormalization())
+        self.model.add(LSTM(256, return_sequences=False, stateful=False, dropout=0.5))
+        self.model.add(BatchNormalization())
+        self.model.add(Dense(output_dim))
+        self.model.add(Activation('sigmoid'))
+
+        self.model.compile(optimizer=SGD(lr=lr),loss='mse')
+        self.prob = None
+
+    def reset(self):
+        self.prob =None
+
+    def predict(self, sample):
+        self.prob = self.model.predict(np.array(sample).reshape((1,-1,self.input_dim)))[0]
+        return self.prob
+
+    def train_on_batch(self,x,y):
+        return self.model.train_on_batch(x,y)
+
+    def save_model(self,model_path):
+        if model_path is not None and self.model is not None:
+            self.model.save_weights(model_path,overwrite=True)
+
+    def load_model(self,model_path):
+        if model_path is not None:
+            self.model.load_weights(model_path)
+
+
